requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
cryptography==44.0.0
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.0
mammoth==1.9.0
markdownify==0.14.1
markitdown==0.0.1a3
MarkupSafe==3.0.2
numpy==2.2.1
openai==1.59.7
openpyxl==3.1.5
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20240706
pillow==11.1.0
puremagic==1.28
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.13.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
XlsxWriter==3.2.0
youtube-transcript-api==0.6.3
